move schema package to x directory
